requirements.txt
aws-cdk-lib==2.68.0
constructs>=10.0.0,<11.0.0
pyspark==3.3.2
tldextract==3.4.0
boto3==1.26.99
pytest==6.2.5
pyclean==2.7.0